enhance singUP process

+++ src/server/server.js
@@ -35,7 +35,7 @@ app.get('/keys', (req, res) => {
 		API_USERNAME,
 	});
 });
-
+// read user email and password
 function readUserRegistration() {
 	return new Promise((resolve, reject) => {
 		fs.readFile('src/data/userRegister.json', (err, data) => {
@@ -45,28 +45,38 @@ function readUserRegistration() {
 		});
 	});
 }
-
+// add email and password to data
 app.post('/singup', (req, res) => {
 	console.log('got requests singup');
 	const body = req.body;
 	const user = {};
 	readUserRegistration().then((result) => {
-		user['id'] = result.users.length + 1;
-		user['email'] = body.email;
-		user['password'] = body.password;
-		Object.assign(user, body);
-		result.users.push(user);
-		fs.writeFile(
-			'src/data/userRegister.json',
-			JSON.stringify(result, null, 2),
-			(err) => {
-				if (err) console.log(err);
-			}
-		);
-		res.send('added');
+		const email = body.email;
+		function emailExists(users) {
+			return users.email === email;
+		}
+		const existed = result.users.find(emailExists);
+		if (existed == 'undefined') {
+			user['id'] = result.users.length + 1;
+			user['email'] = body.email;
+			user['password'] = body.password;
+			Object.assign(user, body);
+			result.users.push(user);
+			// write over our existed file
+			fs.writeFile(
+				'src/data/userRegister.json',
+				JSON.stringify(result, null, 2),
+				(err) => {
+					if (err) console.log(err);
+				}
+			);
+			res.send('added');
+		} else {
+			res.send('existed');
+		}
 	});
 });
-
+// send login data for login verify
 app.get('/login', (req, res) => {
 	console.log('got requests login');
 	readUserRegistration().then((result) => {
